refactor: improve error handling and convert JS files to TypeScript
- Enhance error handling in PlatformConnectors with more specific error types
- Convert remaining JavaScript pages to TypeScript
- Clean up HTML structure in page components by using Next.js Head component
- Remove unused files (_app.js and automation.js)
- Update component interfaces and type definitions
