--- pages/_app.js
@@ -1,15 +1,0 @@
-import '../styles/globals.css'
-import Header from '../components/Header';
-import Footer from '../components/Footer';
-
-function MyApp({ Component, pageProps }) {
-  return (
-    <>
-      <Header />
-      <Component {...pageProps} />
-      <Footer />
-    </>
-  );
-}
-
-export default MyApp
