--- pages/automation.js
@@ -1,191 +1,0 @@
-import React from 'react';
-
-const AutomationPage = () => {
-  return (
-    <html>
-      <head>
-        <style>
-          {`
-            body {
-              font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-              line-height: 1.6;
-              color: #333;
-              max-width: 100%;
-              margin: 0;
-              padding: 0;
-            }
-            .container {
-              display: flex;
-              flex-direction: column;
-              gap: 1.5rem;
-              padding: 1.5rem;
-            }
-            .section {
-              background-color: #f8f9fa;
-              border-radius: 8px;
-              padding: 1.5rem;
-              box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-            }
-            .section h2 {
-              margin-top: 0;
-              color: #2c3e50;
-              border-bottom: 2px solid #4a6491;
-              padding-bottom: 0.5rem;
-            }
-            .automation {
-              margin: 2rem 0;
-            }
-            .automation-container {
-              background-color: white;
-              border-radius: 8px;
-              padding: 1.5rem;
-              box-shadow: 0 2px 4px rgba(0, 0, 0, 0.05);
-            }
-            .automation-tools {
-              display: grid;
-              grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-              gap: 1rem;
-              margin: 1.5rem 0;
-            }
-            .tool-card {
-              background-color: #f8f9fa;
-              border-radius: 8px;
-              padding: 1rem;
-              border-top: 4px solid #4a6491;
-            }
-            .tool-card h4 {
-              margin-top: 0;
-              color: #2c3e50;
-            }
-            .tool-card p {
-              margin: 0.5rem 0;
-              font-size: 0.9rem;
-            }
-            .tool-card ul {
-              padding-left: 1.2rem;
-              margin: 0.5rem 0;
-            }
-            .tool-card li {
-              margin-bottom: 0.3rem;
-              font-size: 0.85rem;
-            }
-            .conclusion {
-              background-color: #2c3e50;
-              color: white;
-              padding: 1.5rem;
-              border-radius: 8px;
-              margin-top: 1.5rem;
-            }
-            .conclusion h3 {
-              margin-top: 0;
-              font-size: 1.3rem;
-              border-bottom: 2px solid rgba(255,255,255,0.3);
-              padding-bottom: 0.5rem;
-            }
-            .conclusion p {
-              margin: 1rem 0;
-            }
-            @media (max-width: 768px) {
-              .automation-tools {
-                grid-template-columns: 1fr;
-              }
-            }
-          `}
-        </style>
-      </head>
-      <body>
-        <div className="container">
-          <div className="section">
-            <div className="automation">
-              <h3>Workflow Automation Opportunities</h3>
-              <p>Leverage your system architecture expertise to automate repetitive tasks in your content workflow, improving efficiency and consistency.</p>
-              
-              <div className="automation-container">
-                <div className="automation-tools">
-                  <div className="tool-card">
-                    <h4>Content Transformation Pipeline</h4>
-                    <p>Automate the adaptation of your primary content for different platforms.</p>
-                    <ul>
-                      <li><strong>Input:</strong> Markdown files from Git repository</li>
-                      <li><strong>Processing:</strong> Platform-specific transformations</li>
-                      <li><strong>Output:</strong> Optimized content for each platform</li>
-                    </ul>
-                    <p>Implementation using Node.js scripts that extract sections, apply templates, and format for each target platform. Integrate with your Git workflow using pre-commit hooks.</p>
-                  </div>
-                  
-                  <div className="tool-card">
-                    <h4>Media Asset Generation</h4>
-                    <p>Automatically create and optimize visual assets for each platform.</p>
-                    <ul>
-                      <li><strong>Input:</strong> Source diagrams and images</li>
-                      <li><strong>Processing:</strong> Resizing, format conversion, optimization</li>
-                      <li><strong>Output:</strong> Platform-specific media assets</li>
-                    </ul>
-                    <p>Implement using Sharp for image processing and a configuration file that defines platform requirements. Trigger on image changes in your Git repository.</p>
-                  </div>
-                  
-                  <div className="tool-card">
-                    <h4>Publication Scheduler</h4>
-                    <p>Schedule and coordinate content publication across platforms.</p>
-                    <ul>
-                      <li><strong>Input:</strong> Content calendar with publication dates</li>
-                      <li><strong>Processing:</strong> API integration with platforms</li>
-                      <li><strong>Output:</strong> Scheduled posts and notifications</li>
-                    </ul>
-                    <p>Build using a combination of platform APIs and scheduling tools. Create a central dashboard for monitoring publication status across all channels.</p>
-                  </div>
-                  
-                  <div className="tool-card">
-                    <h4>Performance Analytics Aggregator</h4>
-                    <p>Collect and consolidate performance metrics from all platforms.</p>
-                    <ul>
-                      <li><strong>Input:</strong> Platform APIs and analytics endpoints</li>
-                      <li><strong>Processing:</strong> Data normalization and aggregation</li>
-                      <li><strong>Output:</strong> Unified performance dashboard</li>
-                    </ul>
-                    <p>Implement using a data pipeline that pulls metrics from each platform API, normalizes the data, and presents it in a unified dashboard for analysis.</p>
-                  </div>
-                  
-                  <div className="tool-card">
-                    <h4>Content Quality Validator</h4>
-                    <p>Automatically check content against quality standards before publication.</p>
-                    <ul>
-                      <li><strong>Input:</strong> Draft content from Git repository</li>
-                      <li><strong>Processing:</strong> Automated checks and validations</li>
-                      <li><strong>Output:</strong> Quality report with improvement suggestions</li>
-                    </ul>
-                    <p>Create a CI/CD pipeline that runs checks for readability, SEO optimization, broken links, and technical accuracy. Integrate with your Git workflow as a pre-merge check.</p>
-                  </div>
-                  
-                  <div className="tool-card">
-                    <h4>Engagement Response Assistant</h4>
-                    <p>Help manage and prioritize engagement across platforms.</p>
-                    <ul>
-                      <li><strong>Input:</strong> Comments and messages from all platforms</li>
-                      <li><strong>Processing:</strong> Categorization and prioritization</li>
-                      <li><strong>Output:</strong> Engagement queue with response suggestions</li>
-                    </ul>
-                    <p>Build a system that aggregates engagement from all platforms, categorizes by type (question, feedback, etc.), and suggests response templates based on content context.</p>
-                  </div>
-                </div>
-              </div>
-            </div>
-            
-            <div className="conclusion">
-              <h3>Leveraging Your Technical Excellence</h3>
-              <p>As a system architect with expertise in application versioning, dependency injection, and feature flags, you're uniquely positioned to create a content workflow that embodies the same principles of modularity, extensibility, and maintainability that you apply to your code.</p>
-              
-              <p>By treating your content as a product with clear versioning, dependencies, and feature toggles, you can create a sustainable system that scales with your growing audience while maintaining the technical excellence that defines your work.</p>
-              
-              <p>The workflow and platform analysis outlined in this document provides a framework that balances technical depth with accessibility, allowing you to share your expertise effectively across multiple platforms while maintaining a consistent voice and quality standard.</p>
-              
-              <p>By implementing the suggested automation opportunities, you'll create a content production system that reflects your architectural approach—efficient, scalable, and built for long-term success.</p>
-            </div>
-          </div>
-        </div>
-      </body>
-    </html>
-  );
-};
-
-export default AutomationPage;
