--- pages/content-adaptation.js
@@ -1,361 +1,0 @@
-import React from 'react';
-
-const ContentAdaptationPage = () => {
-  return (
-    <html>
-      <head>
-        <style>
-          {`
-            body {
-              font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-              line-height: 1.6;
-              color: #333;
-              max-width: 100%;
-              margin: 0;
-              padding: 0;
-            }
-            .container {
-              display: flex;
-              flex-direction: column;
-              gap: 1.5rem;
-              padding: 1.5rem;
-            }
-            .section {
-              background-color: #f8f9fa;
-              border-radius: 8px;
-              padding: 1.5rem;
-              box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-            }
-            .section h2 {
-              margin-top: 0;
-              color: #2c3e50;
-              border-bottom: 2px solid #4a6491;
-              padding-bottom: 0.5rem;
-            }
-            .workflow {
-              margin: 1.5rem 0;
-            }
-            .workflow-container {
-              background-color: white;
-              border-radius: 8px;
-              padding: 1.5rem;
-              box-shadow: 0 2px 4px rgba(0, 0, 0, 0.05);
-            }
-            .workflow-diagram {
-              display: flex;
-              flex-direction: column;
-              gap: 2rem;
-              margin: 2rem 0;
-            }
-            .workflow-stage {
-              display: flex;
-              flex-direction: column;
-              gap: 1rem;
-            }
-            .stage-header {
-              background-color: #4a6491;
-              color: white;
-              padding: 1rem;
-              border-radius: 8px;
-              display: flex;
-              align-items: center;
-              gap: 1rem;
-            }
-            .stage-number {
-              background-color: white;
-              color: #4a6491;
-              width: 36px;
-              height: 36px;
-              border-radius: 50%;
-              display: flex;
-              align-items: center;
-              justify-content: center;
-              font-weight: bold;
-              font-size: 1.2rem;
-            }
-            .stage-title {
-              font-size: 1.3rem;
-              font-weight: bold;
-              margin: 0;
-            }
-            .stage-steps {
-              display: grid;
-              grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
-              gap: 1rem;
-            }
-            .step-card {
-              background-color: #f8f9fa;
-              border-radius: 8px;
-              padding: 1rem;
-              border-left: 4px solid #4a6491;
-            }
-            .step-card h4 {
-              margin-top: 0;
-              color: #2c3e50;
-              font-size: 1.1rem;
-              margin-bottom: 0.5rem;
-            }
-            .step-card ul {
-              padding-left: 1.2rem;
-              margin: 0;
-            }
-            .step-card li {
-              margin-bottom: 0.5rem;
-              font-size: 0.9rem;
-            }
-            .step-card .tip {
-              background-color: #e9f7fe;
-              border-left: 4px solid #3498db;
-              padding: 0.5rem;
-              margin-top: 0.5rem;
-              font-size: 0.85rem;
-            }
-            .step-card .tip strong {
-              color: #3498db;
-            }
-            .content-adaptation {
-              margin: 2rem 0;
-            }
-            .adaptation-container {
-              background-color: white;
-              border-radius: 8px;
-              padding: 1.5rem;
-              box-shadow: 0 2px 4px rgba(0, 0, 0, 0.05);
-            }
-            .adaptation-examples {
-              display: grid;
-              grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
-              gap: 1.5rem;
-              margin: 1.5rem 0;
-            }
-            .adaptation-card {
-              background-color: #f8f9fa;
-              border-radius: 8px;
-              padding: 1.5rem;
-            }
-            .adaptation-card h4 {
-              margin-top: 0;
-              color: #2c3e50;
-              display: flex;
-              align-items: center;
-              gap: 0.5rem;
-            }
-            .adaptation-card h4 .platform-badge {
-              background-color: #4a6491;
-              color: white;
-              padding: 0.2rem 0.5rem;
-              border-radius: 4px;
-              font-size: 0.8rem;
-            }
-            .adaptation-card .example {
-              background-color: white;
-              border: 1px solid #e9ecef;
-              border-radius: 4px;
-              padding: 1rem;
-              margin: 1rem 0;
-              font-size: 0.9rem;
-            }
-            .adaptation-card .example .title {
-              font-weight: bold;
-              color: #2c3e50;
-              margin-bottom: 0.5rem;
-            }
-            .adaptation-card .notes {
-              font-size: 0.85rem;
-              color: #666;
-            }
-            @media (max-width: 768px) {
-              .stage-steps, .adaptation-examples {
-                grid-template-columns: 1fr;
-              }
-            }
-          `}
-        </style>
-      </head>
-      <body>
-        <div className="container">
-          <div className="section">
-            <div className="workflow">
-              <div className="workflow-container">
-                <div className="workflow-diagram">
-                  <div className="workflow-stage">
-                    <div className="stage-header">
-                      <div className="stage-number">3</div>
-                      <h3 className="stage-title">Platform Adaptation & Distribution</h3>
-                    </div>
-                    <div className="stage-steps">
-                      <div className="step-card">
-                        <h4>Content Transformation</h4>
-                        <ul>
-                          <li>Adapt primary content for each target platform</li>
-                          <li>Create platform-specific headlines and hooks</li>
-                          <li>Optimize media assets for each platform's requirements</li>
-                          <li>Create platform-appropriate CTAs and engagement prompts</li>
-                        </ul>
-                        <div className="tip">
-                          <strong>Pro Tip:</strong> Create a content transformation template for each platform to ensure consistency in your adaptation process.
-                        </div>
-                      </div>
-                      
-                      <div className="step-card">
-                        <h4>Publication Scheduling</h4>
-                        <ul>
-                          <li>Schedule primary article on phoenixvc.tech</li>
-                          <li>Plan staggered release across secondary platforms</li>
-                          <li>Coordinate cross-promotion between platforms</li>
-                          <li>Schedule follow-up engagement activities</li>
-                        </ul>
-                        <div className="tip">
-                          <strong>Pro Tip:</strong> Use a social media management tool to schedule and coordinate posts across multiple platforms from a single dashboard.
-                        </div>
-                      </div>
-                      
-                      <div className="step-card">
-                        <h4>Engagement & Promotion</h4>
-                        <ul>
-                          <li>Implement active community engagement strategy</li>
-                          <li>Respond to comments and questions across platforms</li>
-                          <li>Share with relevant communities and industry groups</li>
-                          <li>Encourage sharing and discussion</li>
-                        </ul>
-                        <div className="tip">
-                          <strong>Pro Tip:</strong> Create a set of prepared responses to common questions that maintain your voice while saving time.
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                  
-                  <div className="workflow-stage">
-                    <div className="stage-header">
-                      <div className="stage-number">4</div>
-                      <h3 className="stage-title">Analysis & Optimization</h3>
-                    </div>
-                    <div className="stage-steps">
-                      <div className="step-card">
-                        <h4>Performance Tracking</h4>
-                        <ul>
-                          <li>Monitor engagement metrics across platforms</li>
-                          <li>Track referral traffic to primary content</li>
-                          <li>Analyze audience demographics and behavior</li>
-                          <li>Measure against predetermined KPIs</li>
-                        </ul>
-                        <div className="tip">
-                          <strong>Pro Tip:</strong> Create a unified dashboard that pulls metrics from all platforms to get a holistic view of content performance.
-                        </div>
-                      </div>
-                      
-                      <div className="step-card">
-                        <h4>Content Refinement</h4>
-                        <ul>
-                          <li>Update content based on audience feedback</li>
-                          <li>Optimize underperforming elements</li>
-                          <li>Expand on topics generating high engagement</li>
-                          <li>Create follow-up content addressing common questions</li>
-                        </ul>
-                        <div className="tip">
-                          <strong>Pro Tip:</strong> Keep a "content improvement log" where you track all feedback and ideas for future updates.
-                        </div>
-                      </div>
-                      
-                      <div className="step-card">
-                        <h4>Process Improvement</h4>
-                        <ul>
-                          <li>Document lessons learned for each article</li>
-                          <li>Refine workflow based on production experience</li>
-                          <li>Update platform-specific strategies</li>
-                          <li>Incorporate new tools and techniques</li>
-                        </ul>
-                        <div className="tip">
-                          <strong>Pro Tip:</strong> Hold a brief retrospective after each article to identify what worked well and what could be improved.
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-            
-            <div className="content-adaptation">
-              <h3>Content Adaptation Examples</h3>
-              <p>Strategic approaches for adapting your technical content to different platforms while maintaining consistency and quality.</p>
-              
-              <div className="adaptation-container">
-                <div className="adaptation-examples">
-                  <div className="adaptation-card">
-                    <h4><span className="platform-badge">LinkedIn</span> Professional Adaptation</h4>
-                    <div className="example">
-                      <div className="title">Original Article Section:</div>
-                      <p>"The implementation of dependency injection in enterprise applications requires careful consideration of lifecycle management and scope hierarchies."</p>
-                      
-                      <div className="title">LinkedIn Adaptation:</div>
-                      <p>"3 Critical Factors When Implementing DI in Enterprise Systems:</p>
-                      <p>1. Lifecycle Management<br>2. Scope Hierarchies<br>3. Performance Implications</p>
-                      <p>In my latest article, I break down how these factors impact your architecture decisions... [link]"</p>
-                    </div>
-                    <div className="notes">
-                      <p>Key adaptation elements:</p>
-                      <ul>
-                        <li>Transformed into list format for scannability</li>
-                        <li>Added value proposition (3 Critical Factors)</li>
-                        <li>Professional tone maintained but more conversational</li>
-                        <li>Clear call-to-action with link to full article</li>
-                      </ul>
-                    </div>
-                  </div>
-                  
-                  <div className="adaptation-card">
-                    <h4><span className="platform-badge">Twitter</span> Concise Adaptation</h4>
-                    <div className="example">
-                      <div className="title">Original Article Section:</div>
-                      <p>"Feature flags provide a mechanism for deploying code to production while controlling its visibility and activation through configuration rather than deployment."</p>
-                      
-                      <div className="title">Twitter Thread Adaptation:</div>
-                      <p>1/ Feature flags aren't just for A/B testing—they're a deployment strategy that separates code deployment from feature activation.</p>
-                      <p>2/ This means you can deploy code on Tuesday but activate the feature next Monday without additional deployments.</p>
-                      <p>3/ The real power? Gradual rollouts, instant rollbacks, and personalized experiences—all without touching your codebase.</p>
-                      <p>4/ I've detailed 5 implementation patterns in my new article: [link] #DevOps #FeatureFlags</p>
-                    </div>
-                    <div className="notes">
-                      <p>Key adaptation elements:</p>
-                      <ul>
-                        <li>Broken into thread format for engagement</li>
-                        <li>Each tweet provides standalone value</li>
-                        <li>Technical concept explained through benefits</li>
-                        <li>Strategic hashtags for discoverability</li>
-                      </ul>
-                    </div>
-                  </div>
-                  
-                  <div className="adaptation-card">
-                    <h4><span className="platform-badge">Medium</span> Expanded Adaptation</h4>
-                    <div className="example">
-                      <div className="title">Original Article Section:</div>
-                      <p>"When implementing backward compatibility in APIs, versioning strategies must be carefully considered."</p>
-                      
-                      <div className="title">Medium Adaptation:</div>
-                      <p><strong>The Hidden Costs of API Versioning</strong></p>
-                      <p>Last month, our team had to support three different API versions simultaneously. The technical debt was crushing us.</p>
-                      <p>This experience taught me that versioning isn't just a technical decision—it's a business strategy with real implications for your development velocity.</p>
-                      <p>Here's what we learned about balancing backward compatibility with innovation pace...</p>
-                    </div>
-                    <div className="notes">
-                      <p>Key adaptation elements:</p>
-                      <ul>
-                        <li>Added personal narrative and experience</li>
-                        <li>Highlighted business impact beyond technical details</li>
-                        <li>Created emotional hook with "crushing" technical debt</li>
-                        <li>Positioned as a lesson learned rather than pure instruction</li>
-                      </ul>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </body>
-    </html>
-  );
-};
-
-export default ContentAdaptationPage;
