--- pages/index.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import Header from '../components/Header';
-import Hero from '../components/Hero';
-import Footer from '../components/Footer';
-
-const Home = () => {
-  return (
-    <div>
-      <Header />
-      <Hero />
-      <Footer />
-    </div>
-  );
-};
-
-export default Home;
